Changes dn

+++ c-sharp-syllabus-main/csharp-basics/exercises/TypesAndVariables/Casting/Program.cs
@@ -20,11 +20,9 @@ namespace Casting
             string a = "1";
             int b = 2;
             int c = 3;
-            int d = 4;
-            int e = 5;
-
-            //fixme - should be 15 :|
-            int sum = int.Parse(a) + b + c + d + e;
+            double d = 4;
+            float e = 5;
+            int sum = int.Parse(a) + b + c + (int) d + (int) e;
             Console.WriteLine(sum);
         }
 
@@ -35,8 +33,6 @@ namespace Casting
             int c = 3;
             double d = 4.2;
             float e = 5.3f;
-
-            //fixme - should be 15.5 :| 
             var sum = Math.Round((int.Parse(a) + b + c + d + e), 1);
             Console.WriteLine(sum);
         }
